requirements.txt
beautifulsoup4==4.12.3
goose3==3.1.17
protobuf==5.27.0
Requests==2.32.2
streamlit==1.32.2
google-generativeai==0.5.4